VSCP-3677: Cleanup non-latest Backup Snapshots (#2326)

+++ core/orchestrator/activities/backup_activities.go
@@ -3,7 +3,6 @@ package activities
 import (
 	"context"
 	"fmt"
-	"go.temporal.io/sdk/activity"
 	"time"
 
 	googleproxyclient "github.com/vcp-vsa-control-Plane/vsa-control-plane/clients/google-proxy-client"
@@ -20,6 +19,7 @@ import (
 	"github.com/vcp-vsa-control-Plane/vsa-control-plane/utils/errors"
 	"github.com/vcp-vsa-control-Plane/vsa-control-plane/utils/nillable"
 	"github.com/vcp-vsa-control-Plane/vsa-control-plane/workflow_engine/util"
+	"go.temporal.io/sdk/activity"
 	"go.temporal.io/sdk/temporal"
 	"gorm.io/gorm"
 )
@@ -568,9 +568,11 @@ func (a BackupActivity) SnapmirrorGetOrCreate(ctx context.Context, node *models.
 	snapmirror, err = provider.SnapmirrorRelationshipCreate(params, token)
 	if snapmirror != nil {
 		resp := commonparams.SnapmirrorRelationship{UUID: snapmirror.UUID.String()}
-		if snapmirror.Destination != nil && snapmirror.Destination.UUID != nil {
-			resp.DestinationUUID = nillable.ToPointer(snapmirror.Destination.UUID.String())
+		if snapmirror.Destination == nil || snapmirror.Destination.UUID == nil {
+			return nil, vsaerrors.WrapAsTemporalApplicationError(
+				vsaerrors.NewVCPError(vsaerrors.ErrResourceEmptyError, fmt.Errorf("DestinationUUID not found in snapmirror relationship")))
 		}
+		resp.DestinationUUID = nillable.ToPointer(snapmirror.Destination.UUID.String())
 		activity.RecordHeartbeat(ctx, "SnapmirrorGetOrCreate completed")
 		return &resp, nil
 	}
@@ -1003,30 +1005,30 @@ func (a BackupActivity) GetSmDestinationPathActivity(ctx context.Context, backup
 	return GetSmDestinationPath(backupVault, volume)
 }
 
-// CleanupOldAdhocBackupSnapshotsActivity cleans up older adhoc-backup snapshots for a volume, keeping only the latest one
-func (a BackupActivity) CleanupOldAdhocBackupSnapshotsActivity(ctx context.Context, volume *datamodel.Volume, node *models.Node) error {
+// CleanupOldBackupSnapshotsActivity cleans up older backup snapshots for a volume, keeping only the latest one
+func (a BackupActivity) CleanupOldBackupSnapshotsActivity(ctx context.Context, volume *datamodel.Volume, node *models.Node) error {
 	logger := util.GetLogger(ctx)
 
-	// Get all adhoc-backup snapshots for this volume, ordered by creation time (newest first)
+	// Get all backup snapshots for this volume, ordered by creation time (newest first)
 	snapshots, err := a.SE.GetSnapshotsByTypeAndVolumeID(ctx, SnapshotTypeBackup, volume.ID)
 	if err != nil {
-		logger.Errorf("Failed to get adhoc-backup snapshots for volume %s: %v", volume.Name, err)
+		logger.Errorf("Failed to get backup snapshots for volume %s: %v", volume.Name, err)
 		return vsaerrors.WrapAsTemporalApplicationError(err)
 	}
 
 	// If we have more than 1 snapshot, delete the older ones (skip the first one which is the latest)
 	if len(snapshots) > 1 {
-		logger.Infof("Found %d adhoc-backup snapshots for volume %s, cleaning up %d older snapshots",
+		logger.Infof("Found %d backup snapshots for volume %s, cleaning up %d older snapshots",
 			len(snapshots), volume.Name, len(snapshots)-1)
 
 		// Process older snapshots (skip the first one which is the latest)
 		for i := 1; i < len(snapshots); i++ {
 			snapshot := snapshots[i]
-			logger.Infof("Deleting older adhoc-backup snapshot %s for volume %s", snapshot.Name, volume.Name)
+			logger.Infof("Deleting older backup snapshot %s for volume %s", snapshot.Name, volume.Name)
 
 			// Try to delete the snapshot from ONTAP first
 			if snapshot.SnapshotAttributes != nil && snapshot.SnapshotAttributes.ExternalUUID != "" && volume.VolumeAttributes != nil && volume.VolumeAttributes.ExternalUUID != "" {
-				err := a.DeleteBackupSnapshot(ctx, node, snapshot.SnapshotAttributes.ExternalUUID, volume.VolumeAttributes.ExternalUUID)
+				err = a.DeleteBackupSnapshot(ctx, node, snapshot.SnapshotAttributes.ExternalUUID, volume.VolumeAttributes.ExternalUUID)
 				if err != nil {
 					logger.Errorf("Failed to delete snapshot %s from ONTAP: %v", snapshot.Name, err)
 					// Mark snapshot as error state instead of failing the entire operation
@@ -1039,8 +1041,8 @@ func (a BackupActivity) CleanupOldAdhocBackupSnapshotsActivity(ctx context.Conte
 			}
 
 			// Delete the snapshot from database
-			_, err := a.SE.DeleteSnapshot(ctx, snapshot.UUID)
-			if err != nil {
+			_, err = a.SE.DeleteSnapshot(ctx, snapshot.UUID)
+			if err != nil && !errors.IsNotFoundErr(err) {
 				logger.Errorf("Failed to delete snapshot %s from database: %v", snapshot.Name, err)
 				// Mark snapshot as error state instead of failing the entire operation
 				err = a.markSnapshotAsError(ctx, snapshot, fmt.Sprintf("Failed to delete from database: %v", err))
@@ -1050,10 +1052,21 @@ func (a BackupActivity) CleanupOldAdhocBackupSnapshotsActivity(ctx context.Conte
 				continue
 			}
 
-			logger.Infof("Successfully deleted older adhoc-backup snapshot %s for volume %s", snapshot.Name, volume.Name)
+			// Hydrate snapshot deletion to CCFE
+			snapshot.Volume.Pool = volume.Pool
+			location := utils.GetLocation(*snapshot)
+			snapshot.State = models.LifeCycleStateDeleted
+			snapshot.StateDetails = models.LifeCycleStateDeletedDetails
+			err = a.HydrateSnapshotDeletionToCCFEActivity(ctx, snapshot, volume.Name, location, volume.Account.Name)
+			if err != nil {
+				logger.Errorf("Failed to hydrate snapshot deletion to CCFE for snapshot %s: %v", snapshot.Name, err)
+				continue
+			}
+
+			logger.Infof("Successfully deleted older backup snapshot %s for volume %s", snapshot.Name, volume.Name)
 		}
 	} else {
-		logger.Infof("No cleanup needed for volume %s - found %d adhoc-backup snapshots", volume.Name, len(snapshots))
+		logger.Infof("No cleanup needed for volume %s - found %d backup snapshots", volume.Name, len(snapshots))
 	}
 
 	return nil
